chore: Move volume setting to its own file
- We've decided that we're going to also display a slider.
