+++ mobile/src/modules/audio/_screens.tsx
@@ -1,12 +1,9 @@
 import type { StaticScreenProps } from "@react-navigation/native";
 
-import { playbackStore, usePlaybackStore } from "~/stores/Playback/store";
-
 import { ListLayout } from "~/navigation/layouts/ListLayout";
 
-import { SegmentedList } from "~/components/List/Segmented";
-import { Switch } from "~/components/UI/Switch";
 import { PlaybackParameterSettings } from "./_components/PlaybackParameterSettings";
+import { VolumeSettings } from "./_components/VolumeSettings";
 import { EqualizerSettings } from "./equalizer/components/EqualizerSettings";
 import { ReplayGainSettings } from "./replayGain/components/ReplayGainSettings";
 
@@ -17,26 +14,15 @@ function AudioEffectsView({
     params: { showHidden },
   },
 }: Props) {
-  const restoreVolume = usePlaybackStore((s) => s.restoreVolume);
   return (
     <ListLayout>
       <EqualizerSettings />
       {showHidden ? <PlaybackParameterSettings /> : null}
       <ReplayGainSettings />
-      <SegmentedList.Item
-        labelTextKey="feat.playback.extra.restoreVolume"
-        onPress={toggleRestoreVolume}
-        RightElement={<Switch enabled={restoreVolume} />}
-      />
+      <VolumeSettings />
     </ListLayout>
   );
 }
-
-//#region Helpers
-function toggleRestoreVolume() {
-  playbackStore.setState((prev) => ({ restoreVolume: !prev.restoreVolume }));
-}
-//#endregion
 
 const AudioEffectScreenGroup = {
   screenOptions: {
